fix(mcp-server): correct stale ADR-136 refs to ADR-137, dedupe shipped CHANGELOG bullets
The cross-runtime-duplication-of-security-logic ADR was renumbered from
136 to 137 mid-session after colliding with an unrelated concurrent
session's registry-sync-tier-gate ADR (already on main as ADR-136). The
ADR file and index were fixed at the time, but 5 code/test comments and
2 CHANGELOG bullets still cited the old number — now pointing readers
at the wrong ADR entirely.

Also removes a duplicate pair of CHANGELOG bullets (SMI-6228/6240 Fix +
SMI-6229 Docs) that had ended up under both the already-published
v0.7.12 section and [Unreleased] — an instance of the interleaved-
release CHANGELOG-merge hazard tracked in SMI-6254.

Splits local-inventory.path-safety.helpers.ts (joinPath,
isSafePathComponent, isWithinRoot) out of local-inventory.helpers.ts,
discovered over the 500-line cap while touching it for the ADR fix —
re-exported so the one consumer's import is unaffected.

No behavior change; 106/106 affected tests pass.
