+++ packages/mcp-server/src/utils/local-inventory.helpers.ts
@@ -262,7 +262,7 @@ function makeClaudeMdEntry(
  * collision against a since-disabled plugin would resurface as a false
  * positive.
  *
- * SECURITY-RELEVANT DUPLICATION (ADR-136): this function is the TS
+ * SECURITY-RELEVANT DUPLICATION (ADR-137): this function is the TS
  * reference implementation for a native `.mjs` reimplementation at
  * `scripts/lib/mcp-command-guard.plugin-scan.mjs` (SMI-6229) — not a shared
  * import, because that file runs on a `SessionStart` hook path where this
@@ -378,51 +378,14 @@ export function readMtime(filePath: string): number | undefined {
   }
 }
 
-/**
- * Resolve absolute path joining `dir + filename`. Centralized so future
- * portability work (E-ANTI-1 v2) can swap in a relative-to-home derivation.
- */
-export function joinPath(dir: string, filename: string): string {
-  return path.join(dir, filename)
-}
-
-/**
- * Rejects path separators, `.`/`..` segments, and empty strings — an
- * `enabledPlugins` id component is an unvalidated settings-file KEY, not
- * something guaranteed traversal-free upstream of the plugin scanner
- * (SMI-6228 Source 5, cross-provider review finding GPT-5.6-Sol). Moved
- * here from `local-inventory.ts` to keep that file under the 500-line cap.
- */
-export function isSafePathComponent(component: string): boolean {
-  return (
-    component.length > 0 &&
-    component !== '.' &&
-    component !== '..' &&
-    !component.includes('/') &&
-    !component.includes('\\')
-  )
-}
-
-/**
- * True when `candidate`, once resolved to its REAL (symlink-followed) path,
- * is `root` or nested under it. Lexical `path.resolve` doesn't follow
- * symlinks, so a symlinked cache subdirectory could escape a purely-lexical
- * check even though `readdirSync`/`readFileSync` would then genuinely
- * follow it outside (GPT-5.6-Sol review finding). A nonexistent path makes
- * `realpathSync` throw — treated as "not within root" (fail-soft, same as
- * the caller's existing missing-directory skip).
- */
-export function isWithinRoot(root: string, candidate: string): boolean {
-  let resolvedRoot: string
-  let resolvedCandidate: string
-  try {
-    resolvedRoot = fs.realpathSync(root)
-    resolvedCandidate = fs.realpathSync(candidate)
-  } catch {
-    return false
-  }
-  return resolvedCandidate === resolvedRoot || resolvedCandidate.startsWith(resolvedRoot + path.sep)
-}
+// Path-traversal/symlink-escape guards moved to their own file (SMI-6229
+// follow-up) purely to keep this file under the 500-line cap — re-exported
+// here so `local-inventory.ts`'s existing import from this module still works.
+export {
+  joinPath,
+  isSafePathComponent,
+  isWithinRoot,
+} from './local-inventory.path-safety.helpers.js'
 
 /**
  * `parseYamlFrontmatter` returns `string | string[] | undefined` for
